FIA-151: Move reusable test fixtures
