+++ tests/fixtures/etrade_simulator_scenario.py
